Merge All Features to Main

See merge request PCalape/secret-rasta-mini-project!19

+++ src/main/java/com/rastatech/secretrasta/controller/WishesController.java
@@ -1,14 +1,22 @@
 package com.rastatech.secretrasta.controller;
 
-import com.rastatech.secretrasta.dto.*;
-import com.rastatech.secretrasta.model.CommentEntity;
+import com.rastatech.secretrasta.dto.UpdateWishRequest;
+import com.rastatech.secretrasta.dto.WishPageResponse;
+import com.rastatech.secretrasta.dto.WishRequest;
+import com.rastatech.secretrasta.dto.WishResponse;
 import com.rastatech.secretrasta.model.WishEntity;
+import com.rastatech.secretrasta.repository.WishRepository;
 import com.rastatech.secretrasta.service.UserService;
 import com.rastatech.secretrasta.service.WishService;
 import lombok.RequiredArgsConstructor;
 import org.modelmapper.ModelMapper;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
+import org.springframework.http.HttpStatus;
 import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 import javax.validation.Valid;
 import java.util.List;
@@ -33,31 +41,56 @@ public class WishesController {
     }
 
     @GetMapping
-    public List<WishResponse> fetchWishes() {
-        List<WishEntity> wishes = wishService.fetchWishes();
-        return wishes.stream().map(this::mapToWishResponse).collect(Collectors.toList());
+    public List<WishPageResponse> fetchWishes() {
+        Pageable sortedBy = PageRequest.of(0, 10, Sort.by("updatedAt").descending());
+        return wishService.fetchWishes(sortedBy);
     }
 
-    @GetMapping("/{user_id}")
+    @GetMapping("/user/{user_id}")
     public List<WishResponse> fetchWishesByUser(@PathVariable("user_id") Long userId) {
         List<WishEntity> wishes = wishService.fetchWishesByUser(userId);
         return wishes.stream().map(this::mapToWishResponse).collect(Collectors.toList());
     }
 
     @GetMapping("/{wish_id}")
-    public WishResponse fetchWish(@PathVariable("wish_id") Long wishId) {
-        WishEntity wish = wishService.fetchWish(wishId);
-        return mapToWishResponse(wish);
+    public WishPageResponse fetchWish(@PathVariable("wish_id") Long wishId,
+                                      Authentication auth) {
+        String username = (String) auth.getPrincipal();
+        Long userId = userService.fetchUserByUsername(username).getUserId();
+        return wishService.fetchWishWithMoreDetails(wishId, userId);
+    }
+
+    @GetMapping("/granted/{user_id}")
+    public List<WishPageResponse> fetchWishesGrantedByUser(@PathVariable("user_id") Long userId) {
+        Pageable sortedBy = PageRequest.of(0, 10, Sort.by("updatedAt").descending());
+        return wishService.fetchWishesGrantedByUser(userId, sortedBy);
+    }
+
+    @GetMapping("/liked/{user_id}")
+    public List<WishPageResponse> fetchLikedWishes(@PathVariable("user_id") Long userId) {
+        Pageable sortedBy = PageRequest.of(0, 10);
+        return wishService.fetchLikedWishes(userId, sortedBy);
+    }
+
+    @GetMapping("/donated/{user_id}")
+    public List<WishPageResponse> fetchDonatedWishes(@PathVariable("user_id") Long userId) {
+        return wishService.fetchDonatedWishes(userId);
     }
 
     @PutMapping("/{wish_id}")
     public void updateWish(@PathVariable("wish_id") Long wishId,
-                              @Valid @RequestBody UpdateWishRequest wish) {
+                           @Valid @RequestBody UpdateWishRequest wish,
+                           Authentication auth) {
+        if (!auth.getPrincipal().equals(wishService.fetchWish(wishId).getUser().getUsername()))
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
         wishService.updateWish(wishId, wish);
     }
 
     @DeleteMapping("/{wish_id}")
-    public void deleteWish(@PathVariable("wish_id") Long wishId) {
+    public void deleteWish(@PathVariable("wish_id") Long wishId,
+                           Authentication auth) {
+        if (!auth.getPrincipal().equals(wishService.fetchWish(wishId).getUser().getUsername()))
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
         wishService.deleteWish(wishId);
     }
